Completed implementation

+++ Weather App/Model/WeatherModel/WeatherModel.swift
@@ -25,7 +25,7 @@ struct WeatherModel: Codable {
     let visibility: Int?
     let wind: Wind?
     let clouds: Clouds?
-    let dt: Int?
+    let dt: Double?
     let sys: Sys?
     let timezone, id: Int?
     let name: String?
